backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
pydantic==2.8.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
Pillow==10.0.0



